move pie dataset tests into subfolders
